积分兑换，结算页面积分支付展示问题处理。 店铺商品变更店铺商品数量展示同步。

+++ framework/src/main/java/cn/lili/modules/order/cart/entity/vo/PriceDetailVO.java
@@ -27,7 +27,7 @@ public class PriceDetailVO implements Serializable {
     private Double discountPrice;
 
     @ApiModelProperty(value = "支付积分")
-    private Integer payPoint;
+    private Long payPoint;
 
     @ApiModelProperty(value = "最终成交金额")
     private Double finalePrice;
